Work around Documenter.generate() tests on 0.7
Documenter.generate() does not really really work for Pkg3, but this is
just a fix to get the tests to pass.

Also, not sure why the tests were in formats/markdown.jl, but
test/generate.jl seems more appropriate.
