+++ test/runtests.jl
@@ -50,6 +50,9 @@ println("="^50)
 
     # A simple build outside of a Git repository
     include("nongit/tests.jl")
+
+    # Tests for Documenter.generate().
+    include("generate.jl")
 end
 
 # Additional tests
